% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractBracket.R
\name{extractBracket}
\alias{extractBracket}
\title{Extract a best bracket from a simulation result}
\usage{
extractBracket(sim, restrict = TRUE)
}
\arguments{
\item{sim}{The outcome of a simTourney run}

\item{restrict}{Restrict the brakcet to realistic result: if a team wins a
given game, they must also have won all prior games.  Set to FALSE for
impossible, but still very interesting brackets}
}
\value{
a data.table
}
\description{
Given the results of simTourney, this function pulls out a
single tournament result.
}
\details{
This function starts with the most likely winner, and then assumes
they won all prior games.  It then picks the most likely other team in the
championship game, and assumes that team won all prior games.  As such, it
works backwards to determine a single result from a simulation.
}
