% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_data.R
\docType{data}
\name{PomeryRatings}
\alias{PomeryRatings}
\title{Ken Pomery's Ratings}
\description{
Ken Pomery's Ratings
}
\references{
\url{http://kenpom.com/}
}
\author{
Zach Mayer \email{zach.mayer@gmail.com}
}
\keyword{data}
