% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/printableBracket.R
\name{printableBracket}
\alias{printableBracket}
\title{Generate a printable NCAA bracket}
\usage{
printableBracket(tourney)
}
\arguments{
\item{tourney}{The outcome of a simTourney run}
}
\value{
nothing
}
\description{
Given the results of simTourney, this function generates
a pritable bracket
}
\details{
Runs N simulations of the NCAA tournament
}
\note{
No guarantees the format will be correct!
}
\references{
\url{http://citeseerx.ist.psu.edu/viewdoc/download?doi=10.1.1.60.2859&rep=rep1&type=pdf}
}

