% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_data.R
\docType{data}
\name{regular_season_compact_results}
\alias{regular_season_compact_results}
\title{regular_season_compact_results.csv from Kaggle}
\description{
regular_season_compact_results.csv from Kaggle
}
\references{
\url{https://www.kaggle.com/c/march-machine-learning-mania-2016/data}
}
\author{
Zach Mayer \email{zach.mayer@gmail.com}
}
\keyword{data}
