% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_data.R
\docType{data}
\name{sample_submission}
\alias{sample_submission}
\title{sample_submission.csv from Kaggle}
\description{
sample_submission.csv from Kaggle
}
\references{
\url{https://www.kaggle.com/c/march-machine-learning-mania-2016/data}
}
\author{
Zach Mayer \email{zach.mayer@gmail.com}
}
\keyword{data}
