% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_data.R
\docType{data}
\name{sample_submission_seed_2015}
\alias{sample_submission_seed_2015}
\title{Prelim 2015 prediction baseline}
\description{
Prelim 2015 prediction baseline
}
\author{
Zach Mayer \email{zach.mayer@gmail.com}
}
\keyword{data}

