% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_data.R
\docType{data}
\name{sample_submission_women}
\alias{sample_submission_women}
\title{A simple, seed-based submission for the women's tournament}
\description{
A simple, seed-based submission for the women's tournament
}
\author{
Zach Mayer \email{zach.mayer@gmail.com}
}
\keyword{data}
