% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz_data.R
\docType{data}
\name{seed_print_positions}
\alias{seed_print_positions}
\title{Printing positions for each seed in the tournament.  This is for generating
a printable bracket.}
\description{
Printing positions for each seed in the tournament.  This is for generating
a printable bracket.
}
\author{
Zach Mayer \email{zach.mayer@gmail.com}
}
\references{
\url{http://www.kaggle.com/c/march-machine-learning-mania-2015/forums/t/12775/printable-bracket-for-r}
\url{http://www.kaggle.com/c/march-machine-learning-mania-2015/forums/t/12627/simulating-the-tournament}
\url{http://www.kaggle.com/c/march-machine-learning-mania/forums/t/7309/printable-bracket-in-r}
\url{https://github.com/chmullig/marchmania/blob/master/bracket.R}
}
\keyword{data}

