% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simTourney.R
\name{sim_tourney_internal}
\alias{sim_tourney_internal}
\title{Function to do a single tournament simulation}
\usage{
sim_tourney_internal(preds)
}
\arguments{
\item{preds}{the tournament data to use for the simulation}
}
\value{
a data.table
}
\description{
Chooses a random number.  If pred > r, the lower id team wins.
}
\details{
Internal simulation function
}

