% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simTourney.R
\name{simTourney}
\alias{simTourney}
\title{Simulate the NCAA tournament}
\usage{
simTourney(preds, year = 2014)
}
\arguments{
\item{preds}{Predicted outcomes for ALL possible matchups}

\item{year}{The year of the tournament.  Used to lookup the correct slots.}
}
\value{
a data.table
}
\description{
Given a parsed NCAA bracket, simulate the tournament
}
\details{
Runs N simulations of the NCAA tournament
}
\note{
Can be a little slow
}

