% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_data.R
\docType{data}
\name{tourney_seeds}
\alias{tourney_seeds}
\title{tourney_seeds.csv from Kaggle}
\description{
tourney_seeds.csv from Kaggle
}
\references{
\url{https://www.kaggle.com/c/march-machine-learning-mania-2016/data}
}
\author{
Zach Mayer \email{zach.mayer@gmail.com}
}
\keyword{data}
