% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tourney_slots}
\alias{tourney_slots}
\title{tourney_slots.csv from Kaggle}
\description{
tourney_slots.csv from Kaggle
}
\author{
Zach Mayer \email{zach.mayer@gmail.com}
}
\references{
\url{http://www.kaggle.com/c/march-machine-learning-mania-2015/data}
}
\keyword{data}

