% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simTourney.R
\name{walkTourney}
\alias{walkTourney}
\title{"Walk" an NCAA tournament, from round 0 to round 6}
\usage{
walkTourney(preds, year = 2015, upset_bias = 0)
}
\arguments{
\item{preds}{Predicted outcomes for ALL possible matchups}

\item{year}{The year of the tournament.  Used to subset preds, if preds
contain multiple seasons}

\item{upset_bias}{If you want more upsets in your bracket, a little upset
bias will give close games to the underdog.}
}
\value{
a data.table
}
\description{
Given a parsed NCAA bracket, walk forward through each round
to generate a bracket
}
\details{
If your probabilities are transative, there's no reason to
simulate the tournament.  You can just walk forward to generate a bracket
}
\note{
Much faster than simulation
}
