% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_data.R
\docType{data}
\name{all_slots}
\alias{all_slots}
\title{Every possible NCAA matchup, with the associated slot and next slot}
\description{
Every possible NCAA matchup, with the associated slot and next slot
}
\references{
\url{https://www.kaggle.com/c/march-machine-learning-mania-2016/data}
}
\author{
Zach Mayer \email{zach.mayer@gmail.com}
}
\keyword{data}
