% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parseBracket.R
\name{parseBracket}
\alias{parseBracket}
\title{Parse a Kaggle-formatted bracket}
\usage{
parseBracket(f)
}
\arguments{
\item{f}{Path to the Kaggle-formatted bracket csv}
}
\value{
a data.table
}
\description{
Inputs a bracket in kaggle file format and outputs a data.table
}
\details{
Reads the csv using data.table::fread, then uses splitstr to parse.
problems
}
\note{
Will only have 1 row per game, where team_1 is the lower id team
}
\examples{
f <- system.file('kaggle_data/SampleSubmission.csv', package = "kaggleNCAA", mustWork=TRUE)
dat <- parseBracket(f)
head(dat)
}

