% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/printableBracket.R
\name{printableBracket}
\alias{printableBracket}
\title{Generate a printable NCAA bracket}
\usage{
printableBracket(bracket, add_seed = TRUE, add_prob = TRUE)
}
\arguments{
\item{bracket}{A bracket to print off}

\item{add_seed}{Add the team's seed in parenthesis}

\item{add_prob}{Add the team's probability of making it this far}
}
\description{
Given an NCAA tournament bracket (a list of slots and who won
the game) this function will plot the bracket in a way that can be printed
off.
}
\references{
\url{http://www.kaggle.com/c/march-machine-learning-mania-2015/forums/t/12775/printable-bracket-for-r}
\url{http://www.kaggle.com/c/march-machine-learning-mania-2015/forums/t/12627/simulating-the-tournament}
\url{http://www.kaggle.com/c/march-machine-learning-mania/forums/t/7309/printable-bracket-in-r}
\url{https://github.com/chmullig/marchmania/blob/master/bracket.R}
}

