% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/printableBracket.R
\name{printableBracket}
\alias{printableBracket}
\title{Generate a printable NCAA bracket}
\usage{
printableBracket(tourney)
}
\arguments{
\item{tourney}{The outcome of a simTourney run}
}
\description{
Given the results of simTourney, this function generates
a pritable bracket
}
\details{
Runs N simulations of the NCAA tournament
}
\note{
No guarantees the format will be correct!
}
\references{
\url{http://www.kaggle.com/c/march-machine-learning-mania-2015/forums/t/12775/printable-bracket-for-r}
\url{http://www.kaggle.com/c/march-machine-learning-mania-2015/forums/t/12627/simulating-the-tournament}
\url{http://www.kaggle.com/c/march-machine-learning-mania/forums/t/7309/printable-bracket-in-r}
\url{https://github.com/chmullig/marchmania/blob/master/bracket.R}
}

