% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/zzz_data.R
\docType{data}
\name{sample_submission}
\alias{sample_submission}
\title{sample_submission.csv from Kaggle}
\description{
sample_submission.csv from Kaggle
}
\author{
Zach Mayer \email{zach.mayer@gmail.com}
}
\references{
\url{http://www.kaggle.com/c/march-machine-learning-mania-2015/data}
}
\keyword{data}

