% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simTourney.R
\name{simTourney}
\alias{simTourney}
\title{Simulate the NCAA tournament}
\usage{
simTourney(preds, N = 1000, year = 2015, progress = TRUE,
  upset_bias = 0, seed_diff_upset_multiple = 0)
}
\arguments{
\item{preds}{Predicted outcomes for ALL possible matchups}

\item{N}{number of simulations}

\item{year}{The year of the tournament.  Used to subset preds, if preds
contain multiple seasons}

\item{progress}{If TRUE, a progress bar will be printed}

\item{upset_bias}{If you want more upsets in your bracket, a little upset
bias will give close games to the underdog.}

\item{seed_diff_upset_multiple}{How much to multiply the upset bias by, based on the seed difference.}
}
\value{
a data.table
}
\description{
Given a parsed NCAA bracket, simulate the tournament
}
\details{
Runs N simulations of the NCAA tournament
}
\note{
Can be a little slow
}

