% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz_data.R
\docType{data}
\name{teams}
\alias{teams}
\title{teams.csv from Kaggle}
\description{
teams.csv from Kaggle
}
\author{
Zach Mayer \email{zach.mayer@gmail.com}
}
\references{
\url{https://www.kaggle.com/c/march-machine-learning-mania-2016/data}
}
\keyword{data}

